+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -65,7 +65,6 @@ set(SOURCES
     DOM/Timer.cpp
     DOM/Window.cpp
     DOMTreeModel.cpp
-    DOMTreeJSONModel.cpp
     Dump.cpp
     FontCache.cpp
     HTML/AttributeNames.cpp
